nv50,nvc0: s/uint16/uint32 for constant buffer offset
Looks like a thinko, "Hey, constant buffers can be at most 64 KiB
in size, offset can't be larger." But it can, of course.

I think piglit lacks a test for UBO and BindBufferRange that
tests if it actually works.

+++ src/gallium/drivers/nvc0/nvc0_stateobj.h
@@ -38,7 +38,7 @@ struct nvc0_constbuf {
       const void *data;
    } u;
    uint32_t size;
-   uint16_t offset;
+   uint32_t offset;
    boolean user; /* should only be TRUE if u.data is valid and non-NULL */
 };
